tests/e2e-test/requirements.txt
pytest-playwright
pytest-html
python-dotenv
pytest-check
pytest-reporter-html1
py
beautifulsoup4